requirements.txt
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
boto3==1.34.124
botocore==1.34.124
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
cryptography==42.0.8
defusedxml==0.7.1
demucs==4.0.1
dj-rest-auth==4.0.0
Django==5.0.6
django-allauth==0.52.0
djangorestframework==3.14.0
dora_search==0.1.12
einops==0.8.0
filelock==3.14.0
fsspec==2024.6.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
julius==0.2.7
lameenc==1.7.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
openunmix==1.3.0
pandas==2.2.2
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
submitit==1.5.1
sympy==1.12.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.4
treetable==0.2.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1